[WUMO-340] Location 엔티티 수정 (#118)
* Refactor : [WUMO-341] 후보지 위경도 data type 변경
- 변경 전 : Float -> 변경 후 : Double

* Refactor : [WUMO-342] 후보지 Category 프론트와 통일
- 변경 전 : MEAL 등 4개
- 변경 후 : DRINK 등 9개
- 변동 사항 : STUDY 삭제, DRINK -> DRINKING
나머지는 추가됨

* Refactor : [WUMO-343] 후보지 엔티티 변경으로 파생된 코드 변동

* Chore : CI 오류 수정

+++ src/main/java/org/prgrms/wumo/domain/location/dto/request/LocationRegisterRequest.java
@@ -27,11 +27,11 @@ public record LocationRegisterRequest(
 
 		@NotNull(message = "위도는 필수 입력사항입니다.")
 		@Schema(description = "후보지 위도", example = "34.566", requiredMode = Schema.RequiredMode.REQUIRED)
-		Float latitude,
+		Double latitude,
 
 		@NotNull(message = "경도는 필수 입력사항입니다.")
 		@Schema(description = "후보지 경도", example = "123.56", requiredMode = Schema.RequiredMode.REQUIRED)
-		Float longitude,
+		Double longitude,
 
 		@NotBlank(message = "이미지는 필수 입력사항입니다.")
 		@Schema(description = "후보지 이미지", example = "https://~", requiredMode = Schema.RequiredMode.REQUIRED)
